requirements.txt
azure-ai-language-questionanswering==1.1.0
azure-ai-textanalytics==5.3.0
azure-ai-vision-face==1.0.0b2
azure-ai-vision-imageanalysis==1.0.0
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.20.0
azure-keyvault-secrets==4.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
dotenv==0.9.9
fonttools==4.56.0
idna==3.10
isodate==0.7.2
kiwisolver==1.4.8
matplotlib==3.10.1
msal==1.31.1
msal-extensions==1.2.0
numpy==2.2.3
packaging==24.2
pillow==11.1.0
portalocker==2.10.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
secretclient==0.0.0
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
